+++ SwiftyTesseract/SwiftyTesseract/Enums/SwiftyTesseractError.swift
@@ -9,6 +9,7 @@
 enum SwiftyTesseractError: Error {
   case imageConversionError
   
+  static let noLanguagesErrorMessage = "SwiftyTesseract must be initialized with at least one language"
   static let initializationErrorMessage = "Initialization of SwiftyTesseract has failed. " +
   "Check that the tessdata folder has been added to the project as a folder reference " +
   "and contains the correct .traineddata files for the specified engine mode and language(s)."