+++ SwiftyTesseract/SwiftyTesseract/SwiftyTesseract.swift
@@ -14,7 +14,6 @@ typealias TessBaseAPI = OpaquePointer
 typealias TessString = UnsafePointer<Int8>
 typealias Pix = UnsafeMutablePointer<PIX>?
 
-
 /// A class to perform optical character recognition with the open-source Tesseract library
 public class SwiftyTesseract {
   
@@ -69,30 +68,35 @@ public class SwiftyTesseract {
     return String(tesseractString: tesseractVersion)
   }()
   
+  private init(languageString: String,
+               bundle: Bundle = .main,
+               engineMode: EngineMode = .lstmOnly) {
+    
+    setEnvironmentVariable(.tessDataPrefix, value: bundle.pathToTrainedData)
+    
+    guard TessBaseAPIInit2(tesseract,
+                           bundle.pathToTrainedData,
+                           languageString,
+                           TessOcrEngineMode(rawValue: engineMode.rawValue)) == 0
+    else { fatalError(SwiftyTesseractError.initializationErrorMessage) }
+    
+  }
+  
   // MARK: - Initialization
-  /// Creates an instance of SwiftyTesseract. The tessdata folder MUST be
-  /// in your Xcode project as a folder reference (blue folder icon, not yellow) and be named
-  /// "tessdata"
+  /// Creates an instance of SwiftyTesseract using standard RecognitionLanguages. The tessdata
+  /// folder MUST be in your Xcode project as a folder reference (blue folder icon, not yellow)
+  /// and be named "tessdata"
   ///
   /// - Parameters:
   ///   - languages: Languages of the text to be recognized
   ///   - bundle: The bundle that contains the tessdata folder - default is .main
   ///   - engineMode: The tesseract engine mode - default is .lstmOnly
-  public init(languages: [RecognitionLanguage],
+  public convenience init(languages: [RecognitionLanguage],
               bundle: Bundle = .main,
               engineMode: EngineMode = .lstmOnly) {
     
     let stringLanguages = RecognitionLanguage.createLanguageString(from: languages)
-    
-    // Required for Tesseract to access the .traineddata files
-    setEnvironmentVariable(.tessDataPrefix, value: bundle.pathToTrainedData)
-    
-    // Traps to avoid undefined behavior - TessBaseAPI returns 0 upon successful initialization
-    guard TessBaseAPIInit2(tesseract,
-                           bundle.pathToTrainedData,
-                           stringLanguages,
-                           TessOcrEngineMode(rawValue: engineMode.rawValue)) == 0
-    else { fatalError(SwiftyTesseractError.initializationErrorMessage) }
+    self.init(languageString: stringLanguages, bundle: bundle, engineMode: engineMode)
     
   }
   
@@ -108,6 +112,38 @@ public class SwiftyTesseract {
                           engineMode: EngineMode = .lstmOnly) {
     
     self.init(languages: [language], bundle: bundle, engineMode: engineMode)
+  }
+  
+  /// Creates an instance of SwiftyTesseract using CustomLanguages. The tessdata
+  /// folder MUST be in your Xcode project as a folder reference (blue folder icon, not yellow)
+  /// and be named "tessdata"
+  ///
+  /// - Parameters:
+  ///   - customLanguages: The custom languages of the text to be recognized
+  ///   - bundle: The bundle that contains the tessdata folder - default is .main
+  ///   - engineMode: The tesseract engine mode - default is .lstmOnly
+  public convenience init(customLanguages: [CustomLanguage],
+                          bundle: Bundle = .main,
+                          engineMode: EngineMode = .lstmOnly) {
+    
+    let stringLanguages = CustomLanguage.createLanguageString(from: customLanguages)
+    self.init(languageString: stringLanguages, bundle: bundle, engineMode: engineMode)
+  }
+  
+  /// Convenience initializer for creating an instance of SwiftyTesseract with one custom language
+  /// to avoid having to input an array with one value (e.g. [.customData("klingon")])
+  /// for the languages parameter
+  ///
+  /// - Parameters:
+  ///   - language: The language of the text to be recognized
+  ///   - bundle: The bundle that contains the tessdata folder - default is .main
+  ///   - engineMode: The tesseract engine mode - default is .lstmOnly
+  public convenience init(customLanguage: CustomLanguage,
+                          bundle: Bundle = .main,
+                          engineMode: EngineMode = .lstmOnly) {
+    
+    self.init(customLanguages: [customLanguage], bundle: bundle, engineMode: engineMode)
+    
   }
   
   deinit {
@@ -162,7 +198,6 @@ public class SwiftyTesseract {
     
   }
 
-  
   // MARK: - Helper functions
 
   private func createPix(from image: UIImage) throws -> Pix {
@@ -181,4 +216,3 @@ public class SwiftyTesseract {
   }
   
 }
-