+++ SwiftyTesseract/SwiftyTesseract/dependencies/include/tesseract/genericvector.h
@@ -150,7 +150,7 @@ class GenericVector {
   // Clear the array, calling the clear callback function if any.
   // All the owned callbacks are also deleted.
   // If you don't want the callbacks to be deleted, before calling clear, set
-  // the callback to NULL.
+  // the callback to nullptr.
   void clear();
 
   // Delete objects pointed to by data_[i]
@@ -164,7 +164,7 @@ class GenericVector {
   // Read/Write the array to a file. This does _NOT_ read/write the callbacks.
   // The callback given must be permanent since they will be called more than
   // once. The given callback will be deleted at the end.
-  // If the callbacks are NULL, then the data is simply read/written using
+  // If the callbacks are nullptr, then the data is simply read/written using
   // fread (and swapping)/fwrite.
   // Returns false on error or if the callback returns false.
   // DEPRECATED. Use [De]Serialize[Classes] instead.
@@ -346,8 +346,8 @@ class GenericVector {
   // vector are small enough that for efficiency it makes sense
   // to start with a larger initial size.
   static const int kDefaultVectorSize = 4;
-  inT32   size_used_;
-  inT32   size_reserved_;
+  int32_t   size_used_;
+  int32_t   size_reserved_;
   T*    data_;
   TessCallback1<T>* clear_cb_;
   // Mutable because Run method is not const
@@ -368,14 +368,14 @@ typedef bool (*FileWriter)(const GenericVector<char>& data,
 inline bool LoadDataFromFile(const char* filename, GenericVector<char>* data) {
   bool result = false;
   FILE* fp = fopen(filename, "rb");
-  if (fp != NULL) {
+  if (fp != nullptr) {
     fseek(fp, 0, SEEK_END);
     long size = ftell(fp);
     fseek(fp, 0, SEEK_SET);
     // Trying to open a directory on Linux sets size to LONG_MAX. Catch it here.
     if (size > 0 && size < LONG_MAX) {
       data->resize_no_init(size);
-      result = fread(&(*data)[0], 1, size, fp) == size;
+      result = static_cast<long>(fread(&(*data)[0], 1, size, fp)) == size;
     }
     fclose(fp);
   }
@@ -392,7 +392,7 @@ inline bool LoadDataFromFile(const STRING& filename,
 inline bool SaveDataToFile(const GenericVector<char>& data,
                            const STRING& filename) {
   FILE* fp = fopen(filename.string(), "wb");
-  if (fp == NULL) return false;
+  if (fp == nullptr) return false;
   bool result =
       static_cast<int>(fwrite(&data[0], 1, data.size(), fp)) == data.size();
   fclose(fp);
@@ -524,7 +524,7 @@ class PointerVector : public GenericVector<T*> {
   // Clear the array, calling the clear callback function if any.
   // All the owned callbacks are also deleted.
   // If you don't want the callbacks to be deleted, before calling clear, set
-  // the callback to NULL.
+  // the callback to nullptr.
   void clear() {
     GenericVector<T*>::delete_data_pointers();
     GenericVector<T*>::clear();
@@ -536,20 +536,20 @@ class PointerVector : public GenericVector<T*> {
   // normal GenericVector of those.
   // Returns false in case of error.
   bool Serialize(FILE* fp) const {
-    inT32 used = GenericVector<T*>::size_used_;
+    int32_t used = GenericVector<T*>::size_used_;
     if (fwrite(&used, sizeof(used), 1, fp) != 1) return false;
     for (int i = 0; i < used; ++i) {
-      inT8 non_null = GenericVector<T*>::data_[i] != NULL;
+      int8_t non_null = GenericVector<T*>::data_[i] != nullptr;
       if (fwrite(&non_null, sizeof(non_null), 1, fp) != 1) return false;
       if (non_null && !GenericVector<T*>::data_[i]->Serialize(fp)) return false;
     }
     return true;
   }
   bool Serialize(TFile* fp) const {
-    inT32 used = GenericVector<T*>::size_used_;
+    int32_t used = GenericVector<T*>::size_used_;
     if (fp->FWrite(&used, sizeof(used), 1) != 1) return false;
     for (int i = 0; i < used; ++i) {
-      inT8 non_null = GenericVector<T*>::data_[i] != NULL;
+      int8_t non_null = GenericVector<T*>::data_[i] != nullptr;
       if (fp->FWrite(&non_null, sizeof(non_null), 1) != 1) return false;
       if (non_null && !GenericVector<T*>::data_[i]->Serialize(fp)) return false;
     }
@@ -563,15 +563,15 @@ class PointerVector : public GenericVector<T*> {
   // Also needs T::T(), as new T is used in this function.
   // Returns false in case of error.
   bool DeSerialize(bool swap, FILE* fp) {
-    inT32 reserved;
+    int32_t reserved;
     if (fread(&reserved, sizeof(reserved), 1, fp) != 1) return false;
     if (swap) Reverse32(&reserved);
     GenericVector<T*>::reserve(reserved);
     truncate(0);
     for (int i = 0; i < reserved; ++i) {
-      inT8 non_null;
+      int8_t non_null;
       if (fread(&non_null, sizeof(non_null), 1, fp) != 1) return false;
-      T* item = NULL;
+      T* item = nullptr;
       if (non_null) {
         item = new T;
         if (!item->DeSerialize(swap, fp)) {
@@ -581,13 +581,13 @@ class PointerVector : public GenericVector<T*> {
         this->push_back(item);
       } else {
         // Null elements should keep their place in the vector.
-        this->push_back(NULL);
+        this->push_back(nullptr);
       }
     }
     return true;
   }
   bool DeSerialize(TFile* fp) {
-    inT32 reserved;
+    int32_t reserved;
     if (!DeSerializeSize(fp, &reserved)) return false;
     GenericVector<T*>::reserve(reserved);
     truncate(0);
@@ -600,14 +600,14 @@ class PointerVector : public GenericVector<T*> {
   // retain the integrity of the stream, the caller must call some combination
   // of DeSerializeElement and DeSerializeSkip of the exact number returned in
   // *size, assuming a true return.
-  static bool DeSerializeSize(TFile* fp, inT32* size) {
+  static bool DeSerializeSize(TFile* fp, int32_t* size) {
     return fp->FReadEndian(size, sizeof(*size), 1) == 1;
   }
   // Reads and appends to the vector the next element of the serialization.
   bool DeSerializeElement(TFile* fp) {
-    inT8 non_null;
+    int8_t non_null;
     if (fp->FRead(&non_null, sizeof(non_null), 1) != 1) return false;
-    T* item = NULL;
+    T* item = nullptr;
     if (non_null) {
       item = new T;
       if (!item->DeSerialize(fp)) {
@@ -617,13 +617,13 @@ class PointerVector : public GenericVector<T*> {
       this->push_back(item);
     } else {
       // Null elements should keep their place in the vector.
-      this->push_back(NULL);
+      this->push_back(nullptr);
     }
     return true;
   }
   // Skips the next element of the serialization.
   static bool DeSerializeSkip(TFile* fp) {
-    inT8 non_null;
+    int8_t non_null;
     if (fp->FRead(&non_null, sizeof(non_null), 1) != 1) return false;
     if (non_null) {
       if (!T::SkipDeSerialize(fp)) return false;
@@ -770,7 +770,7 @@ T GenericVector<T>::contains_index(int index) const {
 template <typename T>
 int GenericVector<T>::get_index(T object) const {
   for (int i = 0; i < size_used_; ++i) {
-    ASSERT_HOST(compare_cb_ != NULL);
+    ASSERT_HOST(compare_cb_ != nullptr);
     if (compare_cb_->Run(object, data_[i]))
       return i;
   }
@@ -856,30 +856,25 @@ void GenericVector<T>::set_compare_callback(
 template <typename T>
 void GenericVector<T>::clear() {
   if (size_reserved_ > 0) {
-    if (clear_cb_ != NULL)
+    if (clear_cb_ != nullptr)
       for (int i = 0; i < size_used_; ++i)
         clear_cb_->Run(data_[i]);
     delete[] data_;
-    data_ = NULL;
+    data_ = nullptr;
     size_used_ = 0;
     size_reserved_ = 0;
   }
-  if (clear_cb_ != NULL) {
-    delete clear_cb_;
-    clear_cb_ = NULL;
-  }
-  if (compare_cb_ != NULL) {
-    delete compare_cb_;
-    compare_cb_ = NULL;
-  }
+  delete clear_cb_;
+  clear_cb_ = nullptr;
+  delete compare_cb_;
+  compare_cb_ = nullptr;
 }
 
 template <typename T>
 void GenericVector<T>::delete_data_pointers() {
-  for (int i = 0; i < size_used_; ++i)
-    if (data_[i]) {
-      delete data_[i];
-    }
+  for (int i = 0; i < size_used_; ++i) {
+    delete data_[i];
+  }
 }
 
 
@@ -888,7 +883,7 @@ bool GenericVector<T>::write(
     FILE* f, TessResultCallback2<bool, FILE*, T const &>* cb) const {
   if (fwrite(&size_reserved_, sizeof(size_reserved_), 1, f) != 1) return false;
   if (fwrite(&size_used_, sizeof(size_used_), 1, f) != 1) return false;
-  if (cb != NULL) {
+  if (cb != nullptr) {
     for (int i = 0; i < size_used_; ++i) {
       if (!cb->Run(f, data_[i])) {
         delete cb;
@@ -906,11 +901,11 @@ bool GenericVector<T>::write(
 template <typename T>
 bool GenericVector<T>::read(
     tesseract::TFile* f, TessResultCallback2<bool, tesseract::TFile*, T*>* cb) {
-  inT32 reserved;
+  int32_t reserved;
   if (f->FReadEndian(&reserved, sizeof(reserved), 1) != 1) return false;
   reserve(reserved);
   if (f->FReadEndian(&size_used_, sizeof(size_used_), 1) != 1) return false;
-  if (cb != NULL) {
+  if (cb != nullptr) {
     for (int i = 0; i < size_used_; ++i) {
       if (!cb->Run(f, data_ + i)) {
         delete cb;
@@ -947,7 +942,7 @@ bool GenericVector<T>::Serialize(tesseract::TFile* fp) const {
 // If swap is true, assumes a big/little-endian swap is needed.
 template <typename T>
 bool GenericVector<T>::DeSerialize(bool swap, FILE* fp) {
-  inT32 reserved;
+  int32_t reserved;
   if (fread(&reserved, sizeof(reserved), 1, fp) != 1) return false;
   if (swap) Reverse32(&reserved);
   reserve(reserved);
@@ -961,7 +956,7 @@ bool GenericVector<T>::DeSerialize(bool swap, FILE* fp) {
 }
 template <typename T>
 bool GenericVector<T>::DeSerialize(tesseract::TFile* fp) {
-  inT32 reserved;
+  int32_t reserved;
   if (fp->FReadEndian(&reserved, sizeof(reserved), 1) != 1) return false;
   reserve(reserved);
   size_used_ = reserved;
@@ -969,9 +964,9 @@ bool GenericVector<T>::DeSerialize(tesseract::TFile* fp) {
 }
 template <typename T>
 bool GenericVector<T>::SkipDeSerialize(tesseract::TFile* fp) {
-  inT32 reserved;
+  int32_t reserved;
   if (fp->FReadEndian(&reserved, sizeof(reserved), 1) != 1) return false;
-  return fp->FRead(NULL, sizeof(T), reserved) == reserved;
+  return fp->FRead(nullptr, sizeof(T), reserved) == reserved;
 }
 
 // Writes a vector of classes to the given file. Assumes the existence of
@@ -1001,7 +996,7 @@ bool GenericVector<T>::SerializeClasses(tesseract::TFile* fp) const {
 // If swap is true, assumes a big/little-endian swap is needed.
 template <typename T>
 bool GenericVector<T>::DeSerializeClasses(bool swap, FILE* fp) {
-  inT32 reserved;
+  int32_t reserved;
   if (fread(&reserved, sizeof(reserved), 1, fp) != 1) return false;
   if (swap) Reverse32(&reserved);
   T empty;
@@ -1013,7 +1008,7 @@ bool GenericVector<T>::DeSerializeClasses(bool swap, FILE* fp) {
 }
 template <typename T>
 bool GenericVector<T>::DeSerializeClasses(tesseract::TFile* fp) {
-  inT32 reserved;
+  int32_t reserved;
   if (fp->FReadEndian(&reserved, sizeof(reserved), 1) != 1) return false;
   T empty;
   init_to_size(reserved, empty);
@@ -1024,7 +1019,7 @@ bool GenericVector<T>::DeSerializeClasses(tesseract::TFile* fp) {
 }
 template <typename T>
 bool GenericVector<T>::SkipDeSerializeClasses(tesseract::TFile* fp) {
-  inT32 reserved;
+  int32_t reserved;
   if (fp->FReadEndian(&reserved, sizeof(reserved), 1) != 1) return false;
   for (int i = 0; i < reserved; ++i) {
     if (!T::SkipDeSerialize(fp)) return false;
@@ -1042,9 +1037,9 @@ void GenericVector<T>::move(GenericVector<T>* from) {
   this->size_used_ = from->size_used_;
   this->compare_cb_ = from->compare_cb_;
   this->clear_cb_ = from->clear_cb_;
-  from->data_ = NULL;
-  from->clear_cb_ = NULL;
-  from->compare_cb_ = NULL;
+  from->data_ = nullptr;
+  from->clear_cb_ = nullptr;
+  from->compare_cb_ = nullptr;
   from->size_used_ = 0;
   from->size_reserved_ = 0;
 }